// src/constants/enums.ts
export enum ActivationStatus {
  Active = "ACTIVE",
  Inactive = "INACTIVE",
  Suspended = "SUSPENDED",
  Blocked = "BLOCKED",
}

export enum NotificationType {
  messageRequest = "MESSAGE_REQUEST",
  favoredRequest = "FAVORED_REQUEST",
  phoneNumberRequest = "PHONE_NUMBER_REQUEST",
}

export enum PostStatsType {
  year = "YEAR",
  month = "MONTH",
  week = "WEEK",
}

export enum TaskStatus {
  ToDo = "TO_DO",
  InProgress = "IN_PROGRESS",
  Done = "DONE",
}

export enum TaskPriority {
  High = "HIGH",
  Medium = "MEDIUM",
  Low = "LOW",
}